server/requirements.txt
openenv-core>=0.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pandas>=2.0.0
numpy>=1.24.0
httpx>=0.25.0
openai>=1.0.0
gradio>=4.0.0
